+++ includes/Blocks/Block.php
@@ -7,11 +7,11 @@
 
 namespace WPGraphQL\ContentBlocks\Blocks;
 
-use WPGraphQL\ContentBlocks\Field\BlockSupports\Anchor;
 use WPGraphQL\ContentBlocks\Registry\Registry;
 use WPGraphQL\ContentBlocks\Type\Scalar\Scalar;
 use WPGraphQL\ContentBlocks\Utilities\DOMHelpers;
 use WPGraphQL\ContentBlocks\Utilities\WPGraphQLHelpers;
+use WPGraphQL\ContentBlocks\Type\Scalar\Scalar;
 use WPGraphQL\Utils\Utils;
 use WP_Block_Type;
 
@@ -76,7 +76,6 @@ class Block {
 	private function register_block_type() {
 		$this->register_block_attributes_as_fields();
 		$this->register_type();
-		$this->register_block_support_fields();
 	}
 
 	/**
@@ -100,6 +99,7 @@ class Block {
 						__( 'Attributes of the %s Block Type', 'wp-graphql-content-blocks' ),
 						$this->type_name
 					),
+					'interfaces'  => $this->get_block_attributes_interfaces(),
 					'fields'      => $block_attribute_fields,
 				]
 			);
@@ -120,13 +120,6 @@ class Block {
 				]
 			);
 		}//end if
-	}
-
-	/**
-	 * Registers fields for the block supports.
-	 */
-	private function register_block_support_fields(): void {
-		Anchor::register_to_block( $this->block );
 	}
 
 	/**
@@ -198,6 +191,15 @@ class Block {
 	 */
 	private function get_block_interfaces(): array {
 		return $this->block_registry->get_block_interfaces( $this->block->name );
+	}
+
+	/**
+	 * Gets the GraphQL interfaces that should be implemented by the block attributes object.
+	 *
+	 * @return string[]
+	 */
+	private function get_block_attributes_interfaces(): array {
+		return $this->block_registry->get_block_attributes_interfaces( $this->block->name );
 	}
 
 	/**
@@ -284,10 +286,19 @@ class Block {
 
 					break;
 			}//end switch
-
-			// if type is set to integer, get the integer value of the attribute.
-			if ( 'integer' === $attribute_config['type'] ) {
-				$value = intval( $value );
+			// Post processing of return value based on configured type
+			switch ( $attribute_config['type'] ) {
+				case 'integer':
+					$value = intval( $value );
+					break;
+				case 'boolean':
+					// If the value is empty or false return
+					if ( is_null( $value ) || false === $value ) {
+						break;
+					}
+					// Otherwise it's truthy
+					$value = true;
+					break;
 			}
 
 			return $value;