+++ includes/Blocks/CoreImage.php
@@ -7,6 +7,9 @@
 
 namespace WPGraphQL\ContentBlocks\Blocks;
 
+use WP_Block_Type;
+use WPGraphQL\ContentBlocks\Registry\Registry;
+
 /**
  * Class CoreImage
  */
@@ -28,6 +31,42 @@ class CoreImage extends Block {
 			'selector'  => 'img',
 			'source'    => 'attribute',
 			'attribute' => 'src',
-		],
-	];
+		),
+		'width'        => [ 'type' => 'string' ],
+	);
+
+	/**
+	 * Block constructor.
+	 *
+	 * @param \WP_Block_Type                             $block The Block Type.
+	 * @param \WPGraphQL\ContentBlocks\Registry\Registry $block_registry The instance of the WPGraphQL block registry.
+	 */
+	public function __construct( WP_Block_Type $block, Registry $block_registry ) {
+		parent::__construct( $block, $block_registry );
+		register_graphql_field(
+			$this->type_name,
+			'mediaDetails',
+			[
+				'type'        => 'MediaDetails',
+				'description' => sprintf(
+					// translators: %s is the block type name.
+					__( 'Media Details of the %s Block Type', 'wp-graphql-content-blocks' ),
+					$this->type_name
+				),
+				'resolve'     => function ( $block ) {
+					$attrs = $block['attrs'];
+					$id = $attrs['id'] ?? null;
+					if ( $id ) {
+						$media_details = wp_get_attachment_metadata( $id );
+						if ( ! empty( $media_details ) ) {
+							$media_details['ID'] = $id;
+
+							return $media_details;
+						}
+					}
+					return null;
+				},
+			]
+		);
+	}
 }